python-app/requirements.txt
# Core dependencies
flask==3.0.0
gunicorn==21.2.0
python-dotenv==1.0.0

# Google Cloud dependencies
google-cloud-storage==2.10.0
google-cloud-pubsub==2.18.4
google-cloud-logging==3.8.0

# Database
sqlalchemy==2.0.23
pymysql==1.1.0

# API and HTTP
requests==2.31.0
httpx==0.25.2

# Data processing
pandas==2.1.4
numpy==1.26.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development
black==23.12.0
flake8==6.1.0